+++ whats4/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bouman-whats4-grupo10",
+  "name": "whats4",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
