+++ whats4/public/manifest.json
